Fragments, SQL, Edit Text Addition
The app has added fragments which enable a drawer list which allows the
user to switch between panels.

SQLite Database in progress to completion which creates a table to store
the recipes that will be created by the user.

The user now has the ability to enter in text on the Add Recipe fragment
however not able to save it just yet as database not fully incorporated.