Web API Progress and Changes to Adding a Recipe
Searching for recipes using the Web API is working and provides results,
however, in an XML format which is attached to a text view which
displays all the XML changes to JSON is in progress and interpreting the
data.

Add Recipe form has changes where there is now buttons to add
ingredients and steps to the EditText boxes instead of the user manually
typing everything in. Dialog boxes will pop up asking for the details
and once confirmed it will be displayed in the correct text box.

Feedback Fragment is complete with little form asking for user details
and a button which sends the information to a email application of the
user's choice.