My Recipes, Recipe, Custom Adapter
Main Activity changed a drawer handling event to test the layout of the
Recipe Fragment.

My Recipes Fragment has now incorporated a list view which a custom list
adapter has been created to add new entries from the database to the
list view. Passing only the entry title and image which is displayed in
the list view.