MyRecipes, Recipe, SQL
Database been updated to include a new column Image Path which takes
strings.

Home Fragment to include a button which deletes the table within the
database and creates a new fresh one.

Add Recipe now has incorporated Image Path to be saved into the database
and it accepts NULL.

MyRecipes List has now incorporated onClicks when clicked a new fragment
is opened of that specific recipe, data such as the primary key is
passed through which is used in the fragment to search the database for
that entry and retrieve all its data from title to image path.

Recipe this is where a recipe is displayed in full detail, however, the
current issue is that when there is no image path saved it will display
"No Image" fine but when there is an image path, displaying an image is
problematic and will be fixed asap.