MyRecipes, Recipe, Adapter, AddRecipe
MyRecipes and the Custom Adapter has been changed to pass Titles, Images
and Primary Key by ArrayList as the parameter instead of having to copy
the ArrayList into a seperate Array which wastes memory. Also
accomplished using the file path passed through to locate and display
the thumbnail of the image in the list of recipes.

AddRecipe when successfully saved a recipe into the database clearing
all boxes includes resetting the imageView to the ic_launcher