New drawables and Edits to Recipes
All drawables are not white instead of the previous grey.

Users can now select a recipe from their recipe list to view it in more
detail and have the option to make changes to the recipe if needs be and
save it which updates the current recipe within the database. Buttons
are currently unavailable on the edits fragment.

Share button has been put in place but not functioning at the moment.

Many of the commonly used strings in XML files have been inserted into
the strings.xml file