Restructuring of Code
There is now a recipe object class which deals with gets and sets.
My Recipes list is now fixed and displaying the locally stored recipes.
Saving and Editing a recipe works with the recipe class.
You can now edit the image of a recipe and save it and this will update
the recipe accordingly in the database.